+++ app/models/user.rb
@@ -15,6 +15,10 @@ class User < ApplicationRecord
   has_one :pending_order, -> { where(state: :just_made) }, class_name: 'Order'
   has_many :payments, dependent: :nullify
 
+  has_one_attached :medical_certificate, dependent: :destroy
+
+  validates :medical_certificate, presence: true, blob: { content_type: ['image/jpg', 'image/jpeg', 'image/png', 'application/pdf'], max_size: 2.megabytes }
+
   validates_uniqueness_of :email, allow_blank: false, if: :email_changed?
   validates_format_of     :email, with: URI::MailTo::EMAIL_REGEXP, allow_blank: true, if: :email_changed?