+++ db/migrate/20200719145101_add_medical_certificate_expire_at_to_user.rb
@@ -1,0 +1,5 @@
+class AddMedicalCertificateExpireAtToUser < ActiveRecord::Migration[6.0]
+  def change
+    add_column :users, :medical_certificate_expire_at, :date
+  end
+end